Moved misplaced classes to correct locations.
