+++ src/main/java/com/nexmo/client/voice/endpoints/SendDtmfMethod.java
@@ -1,5 +1,27 @@
-package com.nexmo.client.voice;/*
+/*
  * Copyright (c) 2011-2017 Nexmo Inc
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.nexmo.client.voice.endpoints;/*
+ * Copyright (c) 2011-2016 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,24 +45,26 @@ package com.nexmo.client.voice;/*
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.JWTAuthMethod;
-import com.nexmo.client.voice.endpoints.AbstractMethod;
+import com.nexmo.client.voice.DtmfRequest;
+import com.nexmo.client.voice.DtmfResponse;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.entity.StringEntity;
 import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-public class StopTalkMethod extends AbstractMethod<String, TalkResponse> {
-    private static final Log LOG = LogFactory.getLog(StopTalkMethod.class);
+public class SendDtmfMethod extends AbstractMethod<DtmfRequest, DtmfResponse> {
+    private static final Log LOG = LogFactory.getLog(SendDtmfMethod.class);
 
     private static final String DEFAULT_URI = "https://api.nexmo.com/v1/calls/";
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{JWTAuthMethod.class};
     private String uri = DEFAULT_URI;
 
-    public StopTalkMethod(HttpWrapper httpWrapper) {
+    public SendDtmfMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
     }
 
@@ -50,17 +74,18 @@ public class StopTalkMethod extends AbstractMethod<String, TalkResponse> {
     }
 
     @Override
-    public RequestBuilder makeRequest(String uuid) throws NexmoClientException, UnsupportedEncodingException {
-        String uri = this.uri + uuid + "/talk";
-        return RequestBuilder.delete(uri)
-                .setHeader("Content-Type", "application/json");
+    public RequestBuilder makeRequest(DtmfRequest request) throws NexmoClientException, UnsupportedEncodingException {
+        String uri = this.uri + request.getUuid() + "/dtmf";
+        return RequestBuilder.put(uri)
+                .setHeader("Content-Type", "application/json")
+                .setEntity(new StringEntity(request.toJson()));
     }
 
     @Override
-    public TalkResponse parseResponse(HttpResponse response) throws IOException {
+    public DtmfResponse parseResponse(HttpResponse response) throws IOException {
         String json = EntityUtils.toString(response.getEntity());
         LOG.info("Received: " + json);
-        return TalkResponse.fromJson(json);
+        return DtmfResponse.fromJson(json);
     }
 
     public void setUri(String uri) {
