+++ src/test/java/com/nexmo/client/voice/CallRecordPageTest.java
@@ -1,6 +1,5 @@
-package com.nexmo.client;
 /*
- * Copyright (c) 2011-2016 Nexmo Inc
+ * Copyright (c) 2011-2017 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,6 +20,9 @@ package com.nexmo.client;
  * THE SOFTWARE.
  */
 
+package com.nexmo.client.voice;
+
+import com.nexmo.client.NexmoUnexpectedException;
 import com.nexmo.client.voice.CallRecordPage;
 import org.junit.Before;
 import org.junit.Test;
