+++ src/test/java/com/nexmo/client/voice/SendDtmfMethodTest.java
@@ -3,6 +3,7 @@ package com.nexmo.client.voice;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nexmo.client.HttpWrapper;
+import com.nexmo.client.voice.endpoints.SendDtmfMethod;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
